Web/backend/pyproject.toml
[tool.poetry]
name = "vernify-backend"
version = "0.1.0"
description = "Vernify LMS Backend - FastAPI"
authors = ["Vernify Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
supabase = "^2.3.0"
httpx = "^0.26.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
celery = {extras = ["redis"], version = "^5.3.0"}
redis = "^5.0.0"
litellm = "^1.80.0"
tenacity = "^8.2.0"
structlog = "^24.1.0"
python-frontmatter = "^1.0.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"
psycopg2-binary = "^2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"
black = "^24.1.0"
isort = "^5.13.0"
mypy = "^1.8.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
